+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * use test.log instead of new two dirs
   * output validation details
   * new cover report
+  * smart skip
 
 ### 3.0.2 - 2022-01-25